add support for binstubs
In order to allow this, I end up making a handful of changes.  The
s:bundle_exec variable becomes the s:test_runner_prefix variable, which
gets set in a function.  This simplifies a bit of logic, and lets us set
the variable based on a few (sometimes independent) preferences: whether
to use `bundle exec` or not and whether to use binstubs (and where they
are located).  A few new variables must be added to make this work: by
default binstubs are turned off (though a path is set if you turn them
on), and bundle exec is turned on.