make main functions take optional arguments
This does two things:

1. It preserves the current api: all current mappings still work.
2. It means we can use the same command for custom runners--just pass an
argument to it.

Note that to do this I need to use custom Merge() to append line
numbers.

I was running into a bug where once you ran the nearest test using a
custom command you were basically stuck running the nearest test every
time.  The problem, it turns out, was that the `extend()` method for
dictionaries updates the dictionary in place (makes sense) and that was
somehow 'sticking' the argument list for subsequent calls.  To be
perfectly honest, I don't really understand why this was
happening---something to do with variable scope, I'm guessing, but I do
know that these changes fix it! :)